// src/src/types/types.ts
export enum EBookingDriverState {
  Considering = 1,
  Canceled,
  Performer,
  Arrived,
  Started,
  Finished
}
// ...
export interface IStaticMarker {
  latitude: number,
  longitude: number,
  popup?: string,
  tooltip?: string
}

export interface IAddressPoint {
  address?: string,
  date?: string,
  time?: string
  shortAddress?: string,
  latitude?: number,
  longitude?: number
}

export interface ILoadedAddressPoint extends IAddressPoint {
  address: string
  latitude: number
  longitude: number
  details: IAddressDetails
}

export enum EStatuses {
  Default,
  Loading,
  Success,
  Fail,
  Whatsapp,
  Warning,
}

export interface IRouteInfo {
  /** KM */
  distance: number,
  /** Summary */
  time: {
    hours: number,
    minutes: number
  },
  /** Latitude, longitude */
  points: Array<[number, number]>
}